feat: implement booking system with models, services, and UI screens for managing reservations and notifications

+++ android/app/src/main/java/com/example/HaliSahaApp/data/services/NotificationService.kt
@@ -85,7 +85,7 @@ object NotificationService {
     private fun scheduleSingle(booking: Booking, bookingId: String, hoursBefore: Int) {
         val calendar = Calendar.getInstance()
         calendar.time = booking.date
-        calendar.set(Calendar.HOUR_OF_DAY, booking.startHour)
+        calendar.set(Calendar.HOUR_OF_DAY, booking.startHour.toInt())
         calendar.set(Calendar.MINUTE, 0)
         calendar.set(Calendar.SECOND, 0)
